mount: fix mounting non-root shared mounts
The widest mount of a shared group has to be mounted first and
then other mounts has to be bind-mounted from the first one to
inherite its group.

This patch adds a check in can_mount_now() to mount the widest group
mount first and sets it as a source mount for other members.