Upgrade a travis VM to Ubuntu 16.04
"If the mountain won't come to Muhammad then Muhammad must go to the
mountain."

Here we upgrade a travis VM to Ubuntu 16.04, then checkpoint a travis
process, reboot the system and restore a travis process back. The travis
process doesn't notices that a virtual machine has been rebooted and
continue working...

It is not for production use, it is just an example how criu can be
used in such cases.

Have fan!!!

https://github.com/travis-ci/travis-ci/issues/5821