kerndat: check whether a kernel supports netns id-s or not
Each networn namespaces has a list of ID-s for other namespaces.
So if we request infomation about a veth device, we get an id
for a namespace of a peer device.

These ID-s can be set by users or by kernel when they are required.
CRIU has to restore these ID-s for network namespaces. We have to
remember that one netns can have different id-s in different network
namespaces.