netlink: allow to dump netlink sockets with in-progress operations
Netlink sockets can have in-progress operations and in this case
only a part of information are in a socket queue and other portions
will be genereated as soon as the queue data will be reed.

But we can ask kernel to queue all data. When we call recv() with
MSG_PEEK, the kernel gives us currently queued data and generates
and queues the next portion. Actually criu already dumps all
required data if they fit into the socket buffer.