lazy-pages: add handling of UFFD_EVENT_MADVDONTNEED
When the restored process calls madvise(MADV_DONTNEED), the memory range
specified by the madvise() call should be remapped to zero pfn and we
should stop monitoring this range in order to avoid its pollution with data
the process does not expect.
All we need to do here, is to unregister the memory range from userfaultfd
and the kernel will take care of the rest.

Signed-off-by: Mike Rapoport <rppt@linux.vnet.ibm.com>