lazy-pages: add handling of UFFD_EVENT_REMAP
When the restored process calls mremap(), we will see #PF's on the new
addresses and we have to create a correspondence between the addresses
found in the dump and the actual addresses the process uses. If the
mremap() call causes the mapping to grow, the additional part will receive
zero pages, as expected.

FIXME: is the mapping shrinks, we will try to fill the dropped part and
lazy-pages daemon will fail. It should be possible to track VMA changes in
the lazy-pages daemon, but simpler, and, apparently, more correct would be
to add UFFD_EVENT_MUNMAP to the kernel.

Signed-off-by: Mike Rapoport <rppt@linux.vnet.ibm.com>