page-read: introduce dup_page_read
The dup_page_read performs a shallow copy of a page_read object. It is
required for implementation of fork event in lazy-pages daemon.
When a restored process fork()'s a child, the lazy-pages daemon will handle
page faults of the child process, and it will use the parent process memory
dump for that.

Signed-off-by: Mike Rapoport <rppt@linux.vnet.ibm.com>