tty: check that a master of a bind-mounted slave is in the root mntns
Here we check that a master of a bind-mounted slave was opened in the
root mount namespace. The problem is that we restore all mounts in the
root mount namespace. Only when all mounts are restored, we create other
mount namespaces.  So when we are restoring mounts, we can open files
only in the root mount namespace.

Signed-off-by: Andrei Vagin <avagin@virtuozzo.com>