ns: Reserve pid_ns helpers
Task is able to set "/proc/sys/kernel/ns_last_pid" only for
its active pid_ns. So, if we create a multi-level pid task,
we need helpers, which allow to set pids in whole pid hierarhy.

This patch reserves a helper for every pid_ns from free pids
of this ns.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>