files: Add force_master argument to collect_desc_fle()
This functional allows to make a fle a master on
the time of collection. We will use it to add fake
files when we need to do this after add_fake_fds_masters().

This will be used to add second end of socketpair as
a fake fle (as the first end is placed in the right
place, we will force add the second end there).
See next patches.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>