unix: restore sockets on correct mount points
Currently we restore all sockets in the root mount namespace,
it is obviously incorrect in some cases.

In 4.10 kernel, we added the SIOCUNIXFILE ioctl for unix sockets.  This
ioctl opens a file to which a socket is bound and returns a file
descriptor. Then we can read fdinfo for this descriptor and get mnt_id
of a mount point where this socket is bound.

On restore, we can find a mount namespace by mnt_id and bind a socket to
a file on a required mount point.

Signed-off-by: Andrei Vagin <avagin@virtuozzo.com>