criu/files-reg.c: add build-id validation functionality
efi.h: Required for accessing the build-id of .efi files

This adds functions to find, store and compare with the stored build-id.
Since the CLI options have not been added yet, the file size is checked first
and if valid, the build-id is always validated, this will be changed after
the CLI options and the checksum functions are added.

Signed-off-by: Ajay Bharadwaj <ajayrbharadwaj@gmail.com>