criu: Add a new option `on-demand` when restoring a process
Loading all memory pages at once when restoring is a default behaviour
of criu restore. However in some cases, the process wouldn't access
a lot of memory pages after restoring. In this patch, we introduce
a new option `on-demand`, which allows user to restore a process
using `sys_mmap` instead of `sys_preadv` as an on-demand way.

Signed-off-by: Yuhan Yang <yangfisher1@sjtu.edu.cn>