kerndat: check support for statmount, STATMOUNT_BY_FD
statmount() can be used to query for mount using either a unique mnt_id
or a file descriptor with STATMOUNT_BY_FD. We want to use statmount()
with STATMOUNT_BY_FD to add support for c/r of files on unmounted bind
mounts (mounts unmounted with MNT_DETACH). So check for support for
statmount() and STATMOUNT_BY_FD flag.

Signed-off-by: Bhavik Sachdev <b.sachdev1904@gmail.com>